+++ sql-planner/src/lib.rs
@@ -106,7 +106,9 @@ impl PreparedStatement {
 
         // Only DQL and DML is cached, because it is on the hot path
         // other types of queries are much less likely to be queried again
-        if new_plan.is_dql_or_dml()? {
+        //
+        // EXPLAIN (RAW) queries contain DQL and DML, we don't want them to be cached
+        if new_plan.is_dql_or_dml()? && !new_plan.is_raw_explain() {
             cache.put(cache_key, new_plan.clone())?;
         }